Split behavior parent diagnostic tests
